fix regression introduced in f48fc8e
One detail of `FallingPiece::cells` that I used in the move finder is that the order of the cells for the same shape is the same. This has not been true since switching to SRS true rotation, so placement locations have not been deduplicated correctly. One symptom of this was weird paths involving double-rotated I/S/Z pieces. Another symptom that is more difficult to see was duplicated work in the evaluator, since it got a duplicate placement for some moves.