Add C API to get plans of CC
Add cc_poll_next_move_with_plans and cc_block_next_move_with_plans
C api to get the plans along with the next move.

Caller is responsible for allocating space for the plans and passing
correct size of array. The actually number of plans is returned in
the same parameter.